Process handler that monitors devices
Currently, running an Android app or tests results in the process
handler quitting as soon as the app is launched. Since we are moving
towards a model where the IDE keeps track of the running applications,
this CL introduces a MultiProcessHandler class that keeps track of
running applications on multiple devices. Currently, a few policy
decisions are hard coded in the handler:
 - it detaches from the process by default (doesn't kill it).
 - if a launch happened on multiple devices, then it keeps running as
   long as the app is active on any one of those devices.

There are a bunch of scenarios that are not well handled right now, with
perhaps the major issue being that the handler doesn't know exactly when
the app launched on the device, and as a result is liable to get out of
sync if there was an error while launching..

Change-Id: I7e40dfe8c37745fab5d60513441a9e30f1ced9b8