New wizard framework based on using data models
This new wizard framework will eventually completely replace the
dynamic wizard framework. The major difference between these
frameworks is that steps now interact with explicit data models,
instead of a dynamic hashtable full of miscellaneous key / value
pairs.

The dynamic wizard framework was an attempt to work around the
limitations of IntelliJ's own wizard framework, particularly
needing to enable/disable followup steps dynamically based on choices
made in earlier steps.

Unfortunately, the dynamic wizard framework is wraught with its own
issues, particularly being an approach so abstract that, while
infinitely flexible, maintainers have issues learning it. The API
often feels like it grew organically, with inconsistent public
methods tacked on as needed, and internals exposed to other classes
in the same package, instead of maintaining encapsulatian. And
because it relies on implicit communication across steps using a
generic hashtable concept, it's far too easy for a code update in one
step to crash another one unintentionally.

Despite that, much of what the dynamic wizard gives you is very
powerful and useful, so this new approach will be designed in a way
that converting UIs over shouldn't be too painful (there will be a
clear path to take), while relying on an explicit data model concept
and refreshing the API in order to remove previous concerns.

This CL also includes tests providing 100% code coverage.

Change-Id: Id70ff95c64a8e912f5bafe107b56cea5d2ef3d81