Support instant run and debug
This CL enables instant actions to both run and debug: While running,
pressing run invokes instant run, and likewise for debug.

When a configuration is launched, AndroidRunConfigBase attempts to first
find if there is a compatible run already present. If so, and if the app
can be patched, then it reuses that descriptor and returns a
PatchDeployState. If there is no compatible run, then a full launch is
performed.

Change-Id: I4a1be5625c2e6740190a9118af4948f3b9c3cd58