Make the list of deploy targets extensible
Currently, a launch configuration has a set of radio buttons
corresponding to 3 deploy targets: show a dialog, emulator or USB
device. This CL converts this UI to be a combobox instead, and allows
other plugins to provide additional targets that will be shown as an
option in the combo. The primary consumer of this is the "Cloud Test
Lab" which will allow deploying tests to a set of devices on the cloud.

The main interface is the DeployTarget extension point. This allows an
extension to either completely control the launch by providing a custom
run profile state, or simply provide a DeviceTarget (corresponding to a
locally attached IDevice) and allow the regular Android launch process
to do its job.

Each extension maintains its own corresponding state, and is allowed to
provide a UI (DeployTargetConfigurable) to allow user configuration of
that state. This UI is displayed in the launch config editor when that
particular DeployTarget is selected in the combo box.

Change-Id: I5d21b074aa7a9f5673caf96f5f4ca69c29a966d8