Compute manifest resource checksums and force coldstart
This CL computes a checksum from the contents of the
manifest file and the resources it references. Resources
referenced from the manifest require an app reinstall
(because they're read out of the APK). Prior to this CL
we'd force a rebuild if the manifest itself was changed,
but that's not enough: if you edit the application name
for example (without touching the manifest), that too
will require a restart. We now use these checksums
and compare the installed checksum with the current
checksum to force a rebuild when manifest resoures have
changed.

b/25212170: string for app_name does not update

Change-Id: If1b01bcdeb3d20c8cf775676bbd5e8cf4c0279ab