Fix caching issues in merged manifest computation
MergedManifestInfo maintains a list of Manifest objects corresponding to
each manifest file. It turns out that sometimes, this list was empty if
the process was invoked from within a progress operation and that
process was cancelled. Since we were setting the cache keys before
setting the cache value, this was ending up in an inconsistent state if
the process was cancelled during the cache value computation.

Change-Id: Ife37422a3b59d7641ac1adda08aa5c0b0c22d6e3