Updated default BindingsManager InvokeStrategy logic
Before, BindingsManager always defaulted to relying on IntelliJ's
application framework for invoking an action later, which works fine
in practice. However, putting together standalone Swing demos which
use properties is also an important use case (in a context where
IntelliJ's application framework returns null), so we now detect that
and revert to Swing's invoke later strategy as a fallback.

Change-Id: Ie24416cb3a287de13761ddf3bc44c6578b8c95fc