Add a new UI for the device chooser
This CL adds a new UI for the device chooser dialog. Currently, this
shows up only when cloud testing is enabled. It will become the default
in a subsequent release.

The new UI allows other plugins to contribute additional UX to the
device chooser. Currently, this is hooked up to show the cloud test
configuration panel if it is enabled.

Change-Id: I73cd483f4f14e58e011223b67f293d76fb76a025