179838: Android Studio do not always accept RC version of Gradle
The regexp for splitting up a filename used regexp groups like this:
    gradle-messaging-2.10.jar
           ~~~~~~~~~ ~~~~
           group1    group2

It would then just parse group two (2.10 above) as the version number.

However, it was greedily including dashes, so here's how it processed
a milestone build:
    gradle-messaging-2.5-rc-1.jar
           ~~~~~~~~~~~~~~~~ ~
           group1           group2

That meant it ended up thinking the version number of the milestone build
was 1, not 2.5, and the comparison failed.

Change-Id: I92c2b6e6450190792c105729e7c019f942d83511