Give 'adb root' some time to recover.
'adb root' causes the adb deamon on the device to restart. This usually
takes some time. Right after invoking 'adb root', device.root() tries to
execute a shell command to check if rooting was successful. If adb has
not finished yet (on my system it "never" has), it throws an exception,
even though it became root just fine. This change tries to check for a
while, as long as exceptions are thrown.

Change-Id: I689d75970131319587126fb1370d9405279e2635