Begin work on the new idea.ui.wizard package
This CL creates a new package to support all wizard UI concepts, so that
ModelWizard (a generic wizard framework) can be styled in a way
specific to Android Studio.

At the same time, there's a lot of code that ended up in idea.npw but
doesn't quite belong there (and is accessed from other packages that
otherwise shouldn't care about npw, like the AVD wizard for example).

This first pass has the foundation of all the code we'll need for
styling ModelWizards and ModelWizardSteps, plus there is some
initial migration of utility classes out of the npw package.

Change-Id: I6e813407e5be555d8b8a1d5b336ce65f4c8e1888