Make Expressions always require at least one target
It's an error to create an expression that doesn't change based on
at least one target observable, since otherwise the expression would
never fire. However, it can be easy to forget to set a target
observable by accident, which would be hard to debug. It's better in
that case to fail fast.

For example, this is fine:
bind(destProperty, new ToUpperExpression(srcProperty) { s -> s.toUpper(); }

but this, while it compiles, is a hard to detect user error:

bind(destProperty, new ToUpperExpression() { s -> s.toUpper(); }

Change-Id: I5449fb9d673e1f9718ff5c80d4a5406c66350248