Remove testing of import-project wizard from GuiTestCase base class
This code runs in over a dozen GuiTestCase subclasses and exhibits flaky
behavior suggesting a race condition, sometimes failing to close the
file-chooser dialog and timing out waiting for the project IDE frame.

If we wish to test the import-project wizard, let's do that on its own
rather than include it in widely reused setup code.

Change-Id: I1d5c380a4426d7481175254b5daed8881dc92b0f