New ModelWizard.Builder class
Before, ModelWizard allowed you to add steps for as long as you
wanted until you called start().

However, this complicated the internals of ModelWizard, which needed
to keep track of started/not started state.

Now, the ModelWizard constructor requires all steps be added up
front, but instead of having the users construct a ModelWizard
directly, we require them to use a builder, instead, which mimics
the old interface of being able to collect steps separately before
instantiating the wizard.

Change-Id: I5884679c6dd80388537dd797ac9f750c2a10c7bc