Fix deadlock when opening the theme editor
The deadlock happens because the table renderer uses the same component
as the editor, and that component requires a write lock to set the text.
So painting the table can potentially cause a deadlock.
We now use a simple label for the renderer so that we will no need a
write lock anymore when simply painting the table.

Bug: http://b.android.com/193673
Change-Id: I7fcddfe251821f285650477a79ef10bb1d8f5297