Make the Expression base class easily instantiatable
Currently, to use an Expression, you'd need to use one of its many
provided subclasses, but that doesn't scale. Now, you can instantiate
an Expression on its own, as long as you provide a get method.

While making this change, removed a handful of redundant cases which
should just rely on creating an expression directly instead.

Change-Id: Ie6552744e87001d32fcc1c3cd5e370a7d9613604