Improve last selected device behavior in device picker
When the "Use last selected device" option is enabled in the device
picker, the ShowChooserTargetProvider will now attempt to retrieve the
past selection, see if it is still valid, and if so, use that instead of
prompting again.

There is one scenario where this doesn't work well: If an AVD is
selected to be launched, then the next time around, it will actually be
a running device. In such a case, currently the dialog will pop up
again.

Change-Id: I5d12c0abf7bd505b8afbf6aefa772b879a04441c