Fix TableLayout crash when adding layout as children
This addresses two problems with TableLayout. When adding a new TableRow
in a new row, this will avoid nesting the TableRows.
This CL also removes the layout_width and layout_height of the new
children added to TableRows. This helps to avoid the TableLayout bug
that causes a crash when height="match_parent" is used and one of the
first columns is missing.

Bug: http://b.android.com/158489
Bug: http://b.android.com/186314
Change-Id: I909ff2fdc46805ef0180756428048906b51d68d7