25394715: Show link to update gradle version in run configuration
If we detect that the current gradle version is not compatible with
instant run, then we show a hyperlink to "Update Project". The behavior
is almost identical to the version inside InstantRunConfigurable. About
the only difference is that we issue a synchronous call to gradle sync,
without which the run config editor doesn't seem to pick up the gradle
changes (The message bus notifications weren't being delivered, and the
sync only occured after the dialog was closed).

Fixes b/25394715

Change-Id: I08595bb393cb2a20581a7aa63c2eb9b71f1f37a3