25476017: kill run session when attaching to debugger
If a process is first launched using run, then attached to using the
debugger, we end up with two active sessions: the run session (since the
process is still alive), and the new debug session.

And instant run only knows about the first run session, and not the new
debug session, so we have a really wierd UX.

This CL takes the simple solution: if we attach the debugger to an
existing running process, we just terminate any active run sessions for
the same process.

Bug: b/25476017

Change-Id: I08e6da745bbb7e3f607c1f47714ff8eaad1f55a3