Include dead process in sort logic
The logcat device panel sorts all the clients listed by the target
device. However, when a client dies, we manually have been adding it
to the end of the list ourselves. This provides a confusing
experience, though, for users who expect the item to show up in the
sorted list where it used to be when it was alive.

We fix this by adding the dead client to the list first, before we
sort it.

Related issue: https://code.google.com/p/android/issues/detail?id=194104

Change-Id: Idee3e86f569ebffd115b992fbfdbe584f16fd582