Add a method for checking the accelleration status of the machine.
The method will execute the emulator with the "-accel-check"
parameter and report back if a problem was found on the machine.
Some of the problems we should be able to handle in code in Android
Studio (these will have a SolutionCode != NONE).
Other problems will require actions from the user.

Change-Id: I92bc1de381aca9e9453faf1dd6cbdfa19001ebd4