Using NativeAndroidProject model only for pure native modules.
For a hybrid app which contains both java and native code, the
information about the native code is present in both AndroidProject
and NativeAndroidProject. Until we figure out the correct way to
handle that, using the NativeAndroidProject only for the pure native
modules and will be falling back to the information in AndroidProject
model for hybrid modules.

Change-Id: I7a709199a196657704548c51a33d69c669d2302e