Find old session should not have a run config param
After a recent merge, there was some regression in our usage of
the run configuration to find old sessions. This attempts to
undo that.

The method introduced in the merge used the config name to find
the old session. It should be more precise to use our existing
method that uses the configuration id to find the old session.

Change-Id: I73ed11a98c054d0ec3cf42488a4bf5f2b82e9a78
(cherry picked from commit 5765923dbd872a92377402319db099dc4924876e)