TooltipLabel now searches parents for tooltips
In practice, often a component which gets focus may be an
implementation detail for some larger, parent component (which is
where a tooltip gets set on). Therefore, TooltipLabel has to climb
up its ancestry tree to find a tooltip to display.

However, going up the Component tree unbounded may run across
unrelated tooltips eventually, so this CL also offers an API to
bound how high up the tree it will search for.

Change-Id: I4b962ba884459289e9526c9da9b24f7b0230f14b