Remove run config access to find old session
After a recent merge, there was some regression in our usage of
the run configuration to find old sessions. This attempts to
undo that.

The method introduced in the merge used the config name to find
the old session. It should be more precise to use our existing
method that uses the configuration id to find the old session.

Additionally, the new method introduced added a new if
condition. That condition was placed in the existing method
since it makes sense for all situations.

Change-Id: I8d37149430214b750afa460b62db7cc85d5e7ada