Fix issue with validations in ProjectStructureDialog.
Validations of any paths in ProjectStructureDialog is not working.
It seems, we are trying to add the validation listener pretty early
in dialog creation process and the host object is not initialized by
then. I moving the validation listener installation into the
constructor to make it work.

Change-Id: Ib605a361a2d1608b62c6b3c37a4b3f2a20ed9d01
(cherry picked from commit 7c8006b6e3ce536dfd18fc82b365217102077fa1)