Shorten ApiComboBoxItem's name and provide a generic API
The generic tag is not used in this CL but will be made use of later.
Anyway it makes sense to specify the type for the data wrapped by
this class which allows us to avoid messy casting later.

It's argued I should have added types to all current uses of
ApiComboBoxItems, but we'll be migrating a lot of that code in the
future anyway so we'll do it then and not risk making a mistake now.

Change-Id: I908046c6c82159c8faa7efacbf5111975bb5ef47