Implement a mutating TreeModel for the state view.
In the current implementation, the model of the GL state view's tree is
replaced with a new model each time a different atom is selected. This
causes the tree view to loose its state and reset the view to the
default with no expanded nodes. This is not very user friendly as the
user has to expand the tree again and again to see how the GL state
changes. With this change, instead of replacing the tree model, it is
mutated to match the new GL state, firing the corresponding events so
that the view does not loose its state.

Change-Id: I0baea151243ba20fbbe7d140db73f732122b756d