New properties MVVM demo
MVVM (Model View ViewModel) is a UI design architecture which
separates an app into UI, model objects, and logic that binds the
two. This is accomplished by making a model out of properties,
instead of primitive types, and then binding the UI to its various
fields.

In this demo, you can think of the responsibilities broken up as so:

  CorePropertiesDemo.form (UI)
  Person class (Model)
  CorePropertiesDemo class (ViewModel)

Change-Id: I6f9bf9e0bfbfe5072bb7bc41400151f257a47cdb