Warn users if using 2.0.0-alpha1
Now that 2.0.0-alpha2 is available, which fixes a number of critical bugs
in alpha1, make it more obvious to users that there is a newer version
available. On the first build (typically as part of a run, or a post-sync
generate task) the IDE will check whether you're using an obsolete preview
plugin, and if so, prompt you to update.

Change-Id: I4b73c56ab1bb0cfd1f0954f46731a914b2491a86