Don't kill app when relaunching a session
Currently, the multi process handler kills the process for both detach
and destroy methods. This CL makes detach not kill the process, but just
detaches the process handler from the process.

This allows us to not simply terminate the session without terminating
the app itself. Here are the various scenarios:

 hot swap: the existing session remains alive, we just do a hot swap
 cold swap: the existing session remains alive, the process restarts due to
      the way cold swap is implemented
 no instant run: we disconnect the session, and launch a new one. The
      process dies when "pm install" happens as part of the new launch

Fixes b.android.com/195890 and b.android.com/195214

Change-Id: I638d9cdf81bf02c95e2526f8b20be40353d97d96