Extract Validator logic out of StudioWizardStepPanel
In a followup CL, we'll create a panel whose components will
support validations, but it doesn't make sense for it to be a
StudioWizardStepPanel. Therefore, we extract out a ValidatorPanel,
and if later steps want to use it, they should create their own
ValidatorPanel and wrap it with a StudioWizardStepPanel.

Change-Id: I90efa3d7598e45456479563e39ad6f9d83d5a591