Clean up "New Image Asset" and "New Vector Asset" actions
Currently, the actions for creating Asset Studio wizards are using an
inheritance anti-pattern, particularly one action has a bunch of
general logic for creating an image asset dialog (which was probably
there before), and another action inherits from it and punches holes
out of the base class using overrides.

In short, no implementation logic has changed, but we turn this:

           AndroidStudioAssetAction
                       |
               VectorAssetAction

into this:

            AndroidStudioAssetAction
              /                   \
   ImageAssetAction            VectorAssetAction

This cleanup is done in preparation for migrating the asset studio
wizards to use the new ModelWizard system.

Change-Id: I15f92b6660ae8b36711a681eacb611c5fd0b8ad3