Removed redundant icon generation call
Due to a misunderstanding / miscopy of icon generation code, I was
actually generating all notification icons twice. I originally
thought the version meant which set of icons to generate, but
rather the icon generator creates all sets anyways and just uses
the version to tweak options in some cases.

This code seems to act the same as before now but a bit faster, of
course, since it's doing less work.

Change-Id: Ic82780ffa5a06e399c1634b2e0dc53b1679416f8