Fixed bug where JavaModel was not added to Java library modules during sync.
The problem was that the IDE was checking if the module has an AndroidFacet
before setting the JavaModel, which will never happen, because Java library
modules are not Android modules. The correct way to do this check is to
verify that the module has an AndroidGradleFacet. This facet is set to all
modules (Android and Java) that are actually Gradle projects.

Change-Id: I8bc600e88aab870a679682c26f2dacccf6435b9c