Fixing Go To Instance action for hprof viewer.
Now we are able to go to any nested inner classes, both named and
anonymous. This also includes recursively nested inner classes as
well.

This required manual parsing of the class name, looking up the
psiClass of the outer class name, and then recursively visiting the
psi structure using an anonymous java class visitor. Furthermore,
care was taken to try to resolve to the source file, rather than the
decompiled class file.

Error cases (such as class mismatches between the hprof and psi) will
take the user to the last resolved (inncer) class. There could be
issues where the anonymous inner class as reported by the hprof is
completely different the source files in the project (due to the
source files changing, different libraries, and a myriad of other
reasons), but the user should be aware of this, and this shouldn't
impact the most common case.

Unfortunately, there is no reliable way to detect a difference between
the source and the hprof/allocation traces, so this is probably the
best we can go on.

Change-Id: I3c6dac0f008b8942a64c9228198397b4cf121ec8