New Validator.INFO severity and UI refresh
Before, we used red text to indicate errors and yellow text to
indicate warnings. Now, we use normal label text everywhere but use
an appropriate icon to indicate the severity.

Finally, updated the SDK install step to use the new INFO severity
when showing the message, "Please wait for installation to finish"

Change-Id: I696fc0fed96cd930b072c52464c1e314a7c7682c