Use new API check support to enforce explicit and implicit casts
The API check now enforces both implcit and explicit casts, for
both interfaces and super classes.  For example,
android.view.KeyEvent only extends android.view.InputEvent since
API 9. Therefore, while passing a KeyEvent to an API call (or
assigning to a variable etc) which expects an InputEvent will
compile just fine on recent APIs, this can lead to a runtime
crash.

The API database now records all the interface and superclass
changes across API levels, and the ApiDetector now uses this to
check implicit casts in variable declarations and in assignments,
as well as in method call arguments.

Change-Id: I51cfc9bc2d7539a7441729957308c7df4afb60a2