Clean up AssetStudio class responsibilities
These changes are done in preparation for a CL that supports writing
Vector XML icon files to disk. Before the migration, the old version of
AssetStudioAssetGenerator handled everything and became a bit unweildy
to follow, but now as we've been breaking its logic out into many
smaller classes, it seemed best to finalize that before integrating the
generation of Vector XML logic into the system.

In summary:

- Move image icon -> disk logic into AndroidIconGenerator, where it
  more naturally fits now.
- Rename AssetStudioAssetGenerator to AssetStudioGraphicGeneratorContext
  to emphasize its less general purpose role now that
  AndroidIconGenerator has taken over some of its job.
- Remove misc. utility methods out of AndroidGeneratorContext
- Update comments to clarify the changes.

Change-Id: I6439cc48ca8501ed42a2c2e06566e35d6b7e0f32