Don't show compatibility errors if device was offline
The device picker attempts to determine if an application can be
deployed on a device by examining its API version. But if we can't
obtain the API version from the device (usually happens when the
emulator is starting up, or if adb is unauthorized on device), then
there is no point in giving an incorrect message.

This CL changes the error renderer to not display the details if it was
unable to obtain the version number from the device.

https://code.google.com/p/android/issues/detail?id=194518

Change-Id: I5ec310ebf8406a4a08cb7892897a013bd81927be