More precise metric for when instant run is enabled
Currently, Studio reports the Gradle plugin version and whether the
project setting for instant run is enabled. But this requires the
backend to know which versions of Gradle can support instant run.

This CL changes the report to specify that instant run is enabled only
when the project setting is on, and the project uses an appropriate
version of Gradle.

Change-Id: Ia13419f79c71197e1f885d20037d25da19cb501e