Add categories for lint inspections in the IDE
All (~250) lint checks have an associated category and
in some cases a sub category:
* Correctness
* Correctness:Messages
* Security
* Performance
* Usability:Typography
* Usability:Icons
* Usability
* Accessibility
* Internationalization
* Internationalization:Bidirectional Text

However, when integrated in IntelliJ, these were all clustered under a
single Inspections category: "Android Lint".

I've left this alone in the past because there is some value in
understanding the difference between a lint inspection and an IntelliJ
inspection (lint inspections are configured via lintOptions{} in
Gradle, they're running as part of the lint target on your CI server
etc).

However, there are times when you want to explore specific topics; for
example, a while back I was asked how to turn on all the
performance-related lint checks and turn them into errors.

This CL attempts to get the best of both worlds: It creates "sub
categories" in the Inspections panel: you now have

   Android Lint > Correctness
   Android Lint > Correctness > Messages
   Android Lint > Security
and so on.

This also makes several tweaks to the code which checks that all lint
detectors have been registered in the IDE (and which generates the
registration code if not):

- The lint inspection registrations are now sorted case insensitively
  instead of sensitively. This was not done correctly when keys were
  occasionally inserted manually (e.g. it used to be the case that "B
  < a" but some inspections were added using case insensitivity.)

- A few lint checks did not use the canonical key (derived from the
  lint issue id) for the string bundle key.

- By regenerating all the registrations, some display names were
  updated to reflect changes in the lint summaries after the
  inspection was originally registered.

- Updated code which generates list of lint checks that have
  quickfixes (included in lint's HTML reports) to also include lint
  fixes covered by non-lint inspections.

Change-Id: Id23e7a016739721ce6af493907848af93fbabd5e