Split apk install for M
Gradle plugin version 2.0.0-alpha4 supports a few different schemes for
cold swapping changes. This CL adds support in the IDE for supporting
the cold swap scheme as used in devices with API 23 and above.

Before the build, we check the build id on the device to see if it
matches what is present on disk. If the build ids match, then we can do
an incremental build if the app is still running, a full build
otherwise. If the build ids don't match, we need to do a clean build.

Once the build is complete, we check the results of the build to see
whether it can be hotswapped. If so, we just have to update the running
app.

If we cannot do a hotswap, then we perform the updates as specified by
the build-info.xml output by Gradle. This CL only implements the updates
for installing split APKs, which is used only by API 23 and above.

TODO:
 - InstantRunBuildInfo needs to support other artifact types
   (DEX, RELOAD_DEX, RESTART_DEX)
 - We need installers for the DEX and RESTART_DEX types
 - Post install, the build ids and various other things are cached. The
   installers should either use a single base class, or unify this
   properly
 - Various asserts that make this work only on API 23 should be removed

Change-Id: I6bfd04609c0dee317a5922e2c945422e47229d7a