188912: No code inspection
There have been reports that occasionally, lint stops running in a
certain source file. It takes hours to reproduce this. But by
luck I happened to get the IDE into this state, so I could peek
around to see what was going on.

Lint's background analyzer is integrated as an "external annotator",
and it turns out IntelliJ has code which *only* runs external
annotators if there are no "errors" in the file; from ExternalAnnotator:
  @Nullable
  public InitialInfoType collectInformation(@NotNull PsiFile file, @NotNull Editor editor, boolean hasErrors) {
    return hasErrors ? null : collectInformation(file);
  }
In this case, the IDE believed there was an error in the file
(according to
DaemonCodeAnalyzerEx#getFileStatusMap().wasErrorFound(Document)).

This looks like it's the root issue: the IDE had a stale idea that
there was an error found in the document. However, lint is more than
happy to run even when there are errors in the file (in fact, that's
been something I've wanted to fix for a while - the second you have a
syntax error in your file, all the API checks disappear from the file,
and reappear once you fix the syntax error).

Therefore, we'll simply override the collectInformation method to
unconditionally run lint, even in the presence of errors.

Change-Id: I2623c9ee2e09f566fe2d384eb1f991ffc89c6316