178082: NPE in AS inspections flow
When the user cancels an inspection run, various core PSI methods
start throwing exceptions (ProcessCanceledException). These were not
handled properly by the IDE, and in particular, when thrown during the
initial parse of a file (rather than during the more expensive
computation phase) it could lead to the read lock held during PSI
processing to not be released, which would lead to the IDE hanging.

This CL also hooks up the ProcessCanceledException to lint's own
cancel mechanism, such that it bails more quickly.

Change-Id: Ib78d6752b131fa9f7f976d8fa493d10cbca86ffe