Check that APK is installed before checking build timestamps
Build timestamps are saved on /data/local/tmp, and remain there even if
the package itself is uninstalled. So we do an additional check to make
sure that the app is still installed before checking for the build id.

This still leaves out the possibility that you could uninstall and then
re-install an entirely different APK with the same package name from the
command line. So this is a partial fix for
https://code.google.com/p/android/issues/detail?id=198715

Change-Id: If055843474d59e1140cc639c7c8ed1d23d491c91
(cherry picked from commit d0f7c4be7d128f10de55eee894a75ebcf68f0be9)