Fixed the added dependency related quickfixes when test artifacts support are enabled.
The fix works as follow:
When compute the configuration name, we try to get the
focused file and decide if the file is in android test
or unit test. This is not the perfect method, we'll
see if we could improve that late.

Fix: https://code.google.com/p/android/issues/detail?id=202480

Change-Id: Ie39825e5c31c041109626749f7d68e5cd955d45c
(cherry picked from commit baea137d1a1e1cc2eb212755f3d7fcd78572f9e2)