Restart launch after a failed hotswap
If the application dies while gradle is running and we intended to
hotswap, then rather than terminating the launch, attempt to re-run it.

This works by clearing the execution environment of instant run
settings, then calling it to rerun on the current device. On rerun, we
do a full build.

Change-Id: Id3690c523450193dff96dcc0e382788a4826117f