DexDeployTask: Remove unnecessary explicit launch of activity
When a dex file is being pushed, we get a response indicating whether
the activity needs to be launched or not. We had a small hack which then
explicitly launched the activity from the deploy task.

However, this is unnecessary since the DexDeploy task can only be
created in the case of a cold swap, in which case we always do an
activity launch anyway..

Change-Id: I2ac55cb36fbf03642acdf7448bc0e368d04d385e