Allow launching activity after a hotswap
Currently we don't hotswap if the app is in the background. This is
partly because once the hotswap is done, we have to bring the app to the
foreground, otherwise it makes for a confusing user experience.

This CL adds the infrastructure necessary for launching an activity
(which "should" bring the app to the fg) after a hotswap is complete and
indicates that the activity needs to be launched.

Change-Id: I3c9d5c41c7768b60a26917ee6566135da194ddad