Fix AssertionError due to process handler indicating multiple devices
This CL potentially fixes AssertionEr @ AndroidRunConfigurationBase:415
The assertion was making sure that instant run was launched only on a
single device. Using a set instead of a list should fix the issue
assuming that the multiple clients were all pointing to the same device.

The existence of multiple clients for the same process is yet to be
sorted out though.

Change-Id: Id06d34950084a1c1a26b1c877b1c01de8aa7fa71
(cherry picked from commit 0f4b20631405907def6f1df7517b5a9c910ac067)