Instrumentation tests should not be compiled with INSTANT_DEV
Currently, the INSTANT_DEV option is passed to Gradle as long as we are
not running unit tests. This needs to be expanded to include
instrumentation tests as well (i.e. none of the tests should be compiled
with instant run).

This CL makes two changes:
 - The GradleInvokerOptions class is cleaned up to look at the test
   compile type (instead of isUnitTest)
 - The launch configuration now includes an additional method
   (supportsInstantRun). This allows test run configurations to specify
   that they don't support instant run and hence skip the instant run
   specific deploy selection logic.

The apk installer was caching instant run specific data in a non instant
run aware task. The hack still remains, but is tweaked so that it
continues to work as well.

Fixes https://code.google.com/p/android/issues/detail?id=199145

Change-Id: I2cda1159a9101c3004c6d0dea744f123466cc50a
(cherry picked from commit 4805aabe29e84fefbeef97e07375598d6c39fcc8)